% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLogLikR_prediction.R
\name{calcLogLikR_prediction}
\alias{calcLogLikR_prediction}
\title{calcLogLikR_prediction}
\usage{
calcLogLikR_prediction(par, c)
}
\arguments{
\item{par}{parameters A list(mx,mu,omega,beta), The ordinary scale (ttheta)}

\item{c}{data object (returned from prepareC_calibration)}
}
\description{
Likelihood calculation of (evidence) prediction model (uses calibration info) in R
}
