% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateModel.R
\name{calibrateModel}
\alias{calibrateModel}
\title{calibrateModel}
\usage{
calibrateModel(dat, AT = 10, platform = "MPS", locNames = NULL,
  minStuttOccurence = 3, stutterTypes = NULL, createPlots = FALSE,
  runMCMCvalidation = FALSE, mleOptions = list(maxIter = 1000, delta =
  0.1, seed = 1, printLevel = 0, penalty = 1, omegaAdj = 0.5),
  mcmcOptions = list(niter = 1e+05, delta = 0.01), verbose = TRUE)
}
\arguments{
\item{dat}{Datatable with columns (RefName,SampleName,Locus,Allele,Coverage,Dose)}

\item{AT}{Defined analytical threshold used for the analysis (may be defined per marker)}

\item{platform}{Type of platform. Supports "MPS" or "CE".}

\item{locNames}{The order of markers can be specified (otherwise extracted as unique from dataset)}

\item{minStuttOccurence}{Required number of occurence of a particular stutter type (avoiding overfit)}

\item{stutterTypes}{Stutter types. See getStutterTypeRule for supported types}

\item{createPlots}{Whether to create PDF files showing fitted noise model}

\item{runMCMCvalidation}{Boolean of whether to run MCMCvalidation for checking the posterior distributions}

\item{mleOptions}{Option input for optimization}

\item{mcmcOptions}{Option input for MCMC simulation}

\item{verbose}{Whether to print out progress}
}
\value{
Calibration object used for interpreting a new evidence
}
\description{
A wrapper function for performing calibration
}
\details{
Helpfunction to obtain calibration objects ready for evidence predictions
}
\author{
Oyvind Bleka
}
