% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBracket2seq.R
\name{convertBracket2seq}
\alias{convertBracket2seq}
\title{convertBracket2seq}
\usage{
convertBracket2seq(seqs, brksign = c("\\\\[", "\\\\]"))
}
\arguments{
\item{seqs}{Input sequence in bracket format}

\item{brksign}{Bracket format break symbols}
}
\description{
Helpfunction to convert bracket format into full length sequence
}
