% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMotifReps2seq.R
\name{convertMotifReps2seq}
\alias{convertMotifReps2seq}
\title{convertMotifReps2seq}
\usage{
convertMotifReps2seq(motifRepList, brksign = c("[", "]"))
}
\arguments{
\item{motifRepList}{List with motif blocks}

\item{brksign}{Bracket format break symbols}
}
\description{
Helpfunction to convert sequence in motifReps format back to sequence
}
