% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDiagnosticPlots.R
\name{createDiagnosticPlots}
\alias{createDiagnosticPlots}
\title{createDiagnosticPlots}
\usage{
createDiagnosticPlots(obj)
}
\arguments{
\item{obj}{An returned object from the inferMarkerEfficiency function}
}
\description{
A function to plot the diagnostic of the MarkerEfficienc
}
