% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseMCMC.R
\name{diagnoseMCMC}
\alias{diagnoseMCMC}
\title{diagnoseMCMC}
\usage{
diagnoseMCMC(posttheta, showTrace = FALSE, xlim = NULL, MEAN = NULL,
  SD = NULL, fitBeta = FALSE)
}
\arguments{
\item{posttheta}{Matrix with posterior samples (n x p)}

\item{showTrace}{Whether a trace plot of samples should be shown}

\item{xlim}{x-axis limits}

\item{MEAN}{Mean of superimposed distribution}

\item{SD}{SD of superimposed distribution}

\item{fitBeta}{Whether a Beta instead of normal distribution should be superimposed}
}
\description{
Helpfunction to obtain MCMC diagnostic
}
