% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMCMC.R
\name{doMCMC}
\alias{doMCMC}
\title{doMCMC}
\usage{
doMCMC(logLik, th0, Sigma0, niter = NULL, delta = NULL)
}
\arguments{
\item{logLik}{Objective function to perform MCMC on}

\item{th0}{Start value of parameters}

\item{Sigma0}{Assumed covariance matrix of sampler}

\item{niter}{Number of samples to draw}

\item{delta}{Scaling covariance of the sampler (tweaking acceptance rate)}
}
\description{
Helpfunction to do MCMC
}
