% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpareto.R
\name{dpareto}
\alias{dpareto}
\title{dpareto}
\usage{
dpareto(x, alpha, x0, logged = FALSE)
}
\arguments{
\item{x}{Outcome}

\item{alpha}{Model parameter}

\item{x0}{Lower limit of distr}

\item{logged}{Whether logged values should be returned}
}
\description{
Continuous Probability density function  of Pereto
}
\details{
#Used for modeling noise coverage
}
