% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSUMmodel.R
\name{fitSUMmodel}
\alias{fitSUMmodel}
\title{fitSUMmodel}
\usage{
fitSUMmodel(y, x = NULL, niter = 10, delta = 1,
  restrictDeg = FALSE, model = "GA")
}
\arguments{
\item{y}{Vector with sum peak heights (per loci)}

\item{x}{Vector with fragment lengths (base pair) (only provided for degradation)}

\item{niter}{Number of random samples}

\item{delta}{Standard deviation of normal distribution when drawing random startpoints. Default is 1.}

\item{restrictDeg}{Whether to restrict degradation param to be less than 1}

\item{model}{Selected model for the signals (read/peak heights): {"GA"=gamma,"NB"=negative binomial}}
}
\value{
Estimated parameters (mean,coefVar,slopePara)
}
\description{
A function to fit sumPH of the observed signal intensities
}
\details{
This function is used for fitting a regression model (Gamma or Negative binomial)
}
\examples{
\dontrun{
th = c(1000,0.3,0.6)
n = 100 #number of samples
x = seq(10,300,l=n)
x = (x-125)/100
y = rgamma(n,shape=2/th[2]^2*th[3]^x,scale=th[1]*th[2]^2)
plot(x,y)
fitSUMmodel(y,x,model="GA")
fitSUMmodel(y,x,model="NB")
}
}
