% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMPSevidence.R
\name{genMPSevidence}
\alias{genMPSevidence}
\title{genMPSevidence}
\usage{
genMPSevidence(
  calibration,
  NOC,
  popFreq,
  mu = 1000,
  omega = 0.1,
  beta = 1,
  sorted = FALSE,
  mx = NULL,
  refData = NULL,
  nrep = 1,
  kit = NULL,
  platform = "MPS"
)
}
\arguments{
\item{calibration}{An object indicating the fitted calibration model (marker efficiency,stutters,noise) (MLE based)}

\item{NOC}{Number of contributors}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{mu}{Expected peak heights for a het. single contributor allele}

\item{omega}{Coeffecient of variance of peak heights.}

\item{beta}{Coeffecient related to degree of degradation.}

\item{sorted}{Boolean for wheter sorting the contributors with respect to decreasingly mixture proportions.}

\item{mx}{A vector of known mixture proportions. Default is random uniform.}

\item{refData}{A list of given reference profiles given as refData[[i]][[s]]. Default is random from population.}

\item{nrep}{Number of replicates (same contributors) to generate. Default is 1.}

\item{kit}{Kitname for considered kit (shortname). Required for degradation model}

\item{platform}{The platform name that is used}
}
\value{
List with elements theta,samples,refData where theta is the true parameters of the model. samples is a list with samples which for each samples has locus-list elements with list elements adata and hdata
}
\description{
Function for generating replicated mixture samples given same contributors and model parameters.
}
\details{
genDataset samples random mixture peak heights given as gamma(rho*sum(h_k),tau), with h_k as peak height of k-te contributor.
genData conditions on alleles given by refData. Empty references are generated with population frequencies.
}
\examples{
\dontrun{ 
kit = "ForenSeq"
pkg = path.package("MPSproto")
calib =  readRDS(paste0(pkg,"/paper_stutterChar/calibrated_MPSproto.RDS"))
popFreq =  importMPSfreqs(paste0(pkg,"/paper_stutterChar/freqFile_ForenSeqFWbrack_Norway.csv"))[[1]]
gen = genMPSevidence(calib,2,popFreq,mu=1000,omega=0.2,beta=1,kit=kit )
plotMPS(gen$samples,gen$refData,AT=10)
}
}
\author{
Oyvind Bleka
}
