% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplement.R
\name{getComplement}
\alias{getComplement}
\title{getComplement}
\usage{
getComplement(seqs)
}
\arguments{
\item{seqs}{Original sequence}
}
\value{
seqs Reverse complement of the original seqs
}
\description{
Outputs the reverse complement sequence
}
\details{
Helpfunction 1. If marker is positioned in antisense strand, the function
obtains the complement function and subsequently reverts into 5' to 3'
}
\examples{
\dontrun{
getComplement("ATCGATCG")
}
}
\author{
Oyvind Bleka <oyvble.at.hotmail.com>
}
