% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilteredData.R
\name{getFilteredData}
\alias{getFilteredData}
\title{getFilteredData}
\usage{
getFilteredData(dat, platform = "MPS")
}
\arguments{
\item{dat}{Datatable with columns ("SampleName","Locus","Dose")}

\item{platform}{MPS or CE}
}
\description{
Obtain filtered version of data where parental alleleles are not too close
}
\details{
The purpose of this filter is to avoid that two parental alleles interfers with each others in the stutter analysis
We check the distance between the two alleles:
The marker is not used if one is in the back stutter of the other (dist=1)
The CE is tailored to utilize as much data as possible
}
