% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkerEffStat.R
\name{getMarkerEffStat}
\alias{getMarkerEffStat}
\title{getMarkerEffStat}
\usage{
getMarkerEffStat(markerEfficiency, dfAsInput = FALSE, locNames = NULL)
}
\arguments{
\item{markerEfficiency}{Object returned from inferMarkerEfficiency}

\item{dfAsInput}{If markerEfficiency is a dataframe input (optional)}

\item{locNames}{Locus name can be included to override order (optional)}
}
\description{
Obtain marker efficiency statistics (empirical)
}
