% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifReps.R
\name{getMotifReps}
\alias{getMotifReps}
\title{getMotifReps}
\usage{
getMotifReps(block, brksign = c("\\\\[", "\\\\]"))
}
\arguments{
\item{block}{Input block of sequence}

\item{brksign}{Bracket format break symbols}
}
\description{
Helpfunction to Get motifs and repetitions from short bracket format
}
