% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifRepsSequences.R
\name{getMotifRepsSequences}
\alias{getMotifRepsSequences}
\title{getMotifRepsSequences}
\usage{
getMotifRepsSequences(seqs)
}
\arguments{
\item{seqs}{Input sequences in string format}
}
\description{
Helpfunction for obtaining number of motif repeats for each sequence and blocks in each sequence
}
