% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMPSkit.R
\name{getMPSkit}
\alias{getMPSkit}
\title{getMPSkit}
\usage{
getMPSkit(kit = NULL, fileName = "kit.txt", folderName = NULL)
}
\arguments{
\item{kit}{Short name of kit: See supported kits with getKit()}

\item{fileName}{Name of kit file used to extract kitinformation}

\item{folderName}{Name of folder containing the kit file 'fileName'. Using package folder if not specified.}
}
\value{
res A data frame with kit information
}
\description{
Function to get kit information
}
\details{
Returns kit information. Modified by Oyvind Bleka
}
