% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutterData.R
\name{getStutterData}
\alias{getStutterData}
\title{getStutterData}
\usage{
getStutterData(dat, stutterTypes = NULL, platform = "MPS",
  verbose = FALSE)
}
\arguments{
\item{dat}{Datatable with columns (RefName,SampleName,Locus,Allele,Coverage,Dose)}

\item{stutterTypes}{Stutter types. Supports (BW1,BW2,FW1,FW2,DBW1,TBW1,FWBW,BWFW). Also special -2bp stutters for CE}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only consider only some stutter types (but also include -2bp)}

\item{verbose}{Whether to print out progress}
}
\description{
helpfunction to obtain stutter data table used for making inference
}
\details{
The nomenlclature for stutters are as follows:
BW1 means that the LUS motif is providing a backward stutter (n-1).
FW1 means that the LUS motif is providing a forward stutter (n+1).
BW2 means that the non-LUS motif is providing a backward stutter (n-1).
etc... prefix "D" means double (n-2)
n0 stutters are named as either FWBW or BWFW depending on which occuring first in the sequence (hence orientation matters!)
}
