% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutteredSequence.R
\name{getStutteredSequence}
\alias{getStutteredSequence}
\title{getStutteredSequence}
\usage{
getStutteredSequence(seqs, type = "BW1", platform = "MPS")
}
\arguments{
\item{seqs}{Input sequences}

\item{type}{Stutter type. Supports (BW1,BW2,FW1,FW2,DBW1,TBW1,DFW1,FWBW,BWFW)}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only consider only some stutter types (BW1,DBW1,FW1)}
}
\description{
Helpfunction for obtaining different types of stutters
}
