% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutterExpectations.R
\name{getStutterExpectations}
\alias{getStutterExpectations}
\title{getStutterExpectations}
\usage{
getStutterExpectations(c)
}
\arguments{
\item{c}{Output from prepareC_prediction}
}
\value{
A list
}
\description{
Obtain expected stutter proportions
}
\details{
Gives an overview over stutter proportion between each of the sequences
}
\author{
Oyvind Bleka
}
