% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutterIndex.R
\name{getStutterIndex}
\alias{getStutterIndex}
\title{getStutterIndex}
\usage{
getStutterIndex(seqs, type = "BW1", platform = "MPS", locus = NULL)
}
\arguments{
\item{seqs}{Input sequences}

\item{type}{Stutter type. See getStutterTypeRule for supported types}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only LUS based stutter types}

\item{locus}{Locus names (necessary for some markers of CE to obtain useful BLMM)}
}
\description{
Helpfunction for obtaining stutter index between alleles for different types of stutters
}
\details{
The nomenlclature for stutters are as follows:
BW1 means that the LUS motif is providing a backward stutter (n-1).
FW1 means that the LUS motif is providing a forward stutter (n+1).
BW2 means that the non-LUS motif is providing a backward stutter (n-1).
etc... prefix "D" means double (n-2 or n+2)
n0 stutters are named as either FWBW or BWFW depending on which occuring first in the sequence (hence orientation matters!)
}
\examples{
\dontrun{
getStutterIndex(c("[ATCG]10","[ATCG]9"),"BW1")
getStutterIndex(c("[ATCG]10","[ATCG]9"),"BW2")
getStutterIndex(c("[ATCG]9","[ATCG]5 TTGA [ATCG]10","[ATCG]5 TTGA [ATCG]9"),"BW1")
getStutterIndex(c("[ATCG]9","[ATCG]5 TTGA [ATCG]10","[ATCG]5 TTGA [ATCG]9"),"BW2")
getStutterIndex(c("[ATCG]9","[ATCG]5 TTGA [ATCG]10","[ATCG]4 TTGA [ATCG]10"),"BW2")
}
}
