% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutterTypeRule.R
\name{getStutterTypeRule}
\alias{getStutterTypeRule}
\title{getStutterTypeRule}
\usage{
getStutterTypeRule(type, platform = "MPS")
}
\arguments{
\item{type}{Stutter type. Supports (BW1,BW2,FW1,FW2,DBW1,FWBW,BWFW)}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only LUS based stutter types}
}
\description{
Helpfunction to get motifchange for a given stutter type
}
