% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpperLR.R
\name{getUpperLR}
\alias{getUpperLR}
\title{getUpperLR}
\usage{
getUpperLR(mle, scale = TRUE)
}
\arguments{
\item{mle}{Fitted object using inferEvidence function.}

\item{scale}{Whether to scale the RMP when having more than 2 unknowns (recommended)}
}
\value{
Returning upper boundary LR (log10 scale)
}
\description{
Obtaining upper boundary LR for POI
}
\details{
The function takes a fitted MLE object under Hd (where hypothesis$knownRef is the POI)
}
