% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{gui}
\usage{
gui(envirfile = NULL, envir = NULL)
}
\arguments{
\item{envirfile}{A Rdata file including a saved environment of a project}

\item{envir}{A environment object}
}
\description{
gui is a GUI wrapper (gWidgets) for the MPSproto package
}
\author{
Oyvind Bleka
}
