% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMPSfreqs.R
\name{importMPSfreqs}
\alias{importMPSfreqs}
\title{importMPSfreqs}
\usage{
importMPSfreqs(fn)
}
\arguments{
\item{fn}{File name(s).}
}
\value{
A list giving the population frequencies for every population
}
\description{
importMPSfreqs is a function to import allele frequencies directly from file(s) supporting text format
}
\details{
The function reads local text file as input, and return a list of populations.
}
