% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMPSsample.R
\name{importMPSsample}
\alias{importMPSsample}
\title{importMPSsample}
\usage{
importMPSsample(fn)
}
\arguments{
\item{fn}{A file name to import data from (Evid or refs)}
}
\value{
outL A data list
}
\description{
Converting table to list format (helpfunction)
}
