% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferEvidence2.R
\name{inferEvidence2}
\alias{inferEvidence2}
\title{inferEvidence2}
\usage{
inferEvidence2(mlefit, calibration = NULL, priorAsLN = FALSE,
  reOpt = FALSE, nDone = 3, delta = 1, steptol = 1e-04,
  seed = NULL, verbose = FALSE)
}
\arguments{
\item{mlefit}{Output object from inferEvidence function}

\item{calibration}{A list from calibration results (per marker). Must contain the following elements (in following order):
AT Analytical threshold 
markerEff Marker efficiency param (possibly also SD if running inferEvidence2)}

\item{priorAsLN}{Whether model prior of marker efficiency as a log-normal (SD from logged values)}

\item{reOpt}{Whether to re-run optimization with random start values}

\item{nDone}{Number of optimizations required providing equivalent results (same logLik value obtained)}

\item{delta}{Scaling of variation of normal distribution when drawing random startpoints. Default is 1.}

\item{steptol}{Argument used in the nlm function for faster return from the optimization (tradeoff is lower accuracy).}

\item{seed}{The user can set seed if wanted}

\item{verbose}{Whether to print out progress}
}
\description{
Function to perform inference of new evidence (Penalized MLE based)
}
\details{
Optimizes sample specific params (mx,mu,omega, Am). Using MLE as initial params for optimization. Marker efficiency parameter is penalized on prior.
}
