% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferMarkerEfficiency.R
\name{inferMarkerEfficiency}
\alias{inferMarkerEfficiency}
\title{inferMarkerEfficiency}
\usage{
inferMarkerEfficiency(
  dat,
  AT,
  locNames = NULL,
  runMCMC = FALSE,
  mleOptions = list(maxIter = 1000, delta = 0.1, steptol = 1e-06, seed = 1000),
  mcmcOptions = list(niter = 10000, delta = 0.025, seed = 1),
  verbose = FALSE,
  mleObj = NULL
)
}
\arguments{
\item{dat}{Datatable with columns (SampleName,Locus,Coverage)}

\item{AT}{Vector with analytical threshold per marker}

\item{locNames}{The order of markers can be specified (otherwise extracted as unique from dataset)}

\item{runMCMC}{Whether to run MCMC for inferring the posterior distribution}

\item{mleOptions}{Option input for optimization}

\item{mcmcOptions}{Option input for MCMC simulation}

\item{verbose}{Whether to print out progress}

\item{mleObj}{An output from inferMarkerEfficiency with mle fit (used for directly performing MCMC)}
}
\description{
Infer marker efficiency
}
\details{
Using the sum of the coverage per marker as data (per sample), 
the function estimates marker efficiency and sample specific EXP(mu) and CV(omega). 
Using Maximum likelihood estimation for a gamma distribution. 
Requires specification of AT to account for dropout markers.

The function also performs MCMC for diagnostic: the mleObj argument allows for possibility for doing MCMC directly.
}
