% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferNoiseModel.R
\name{inferNoiseModel}
\alias{inferNoiseModel}
\title{inferNoiseModel}
\usage{
inferNoiseModel(sampleNames, locNames, noisedata, AT, print2pdf = NULL)
}
\arguments{
\item{sampleNames}{The samples names used for calibrating the noise model (should also count situations without noise)}

\item{locNames}{The marker names used in calibration.}

\item{noisedata}{A table with potential noise data (not typical stutters). Can be obtained from the ObtainStutterData function (noisetable)}

\item{AT}{A detection threshold value or thresholds per marker. NULL ignores filtering.}

\item{print2pdf}{Name of pdf file for plots}
}
\description{
Fitting noise model in data
}
\details{
Fitting two parts of the noise model: 
1) Fitting the number of noise-observations (per sample) to a geometric distr (1 param) 
2) Fitting the coverage/reads to the pareto distribution
}
