% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferStutterModel.R
\name{inferStutterModel}
\alias{inferStutterModel}
\title{inferStutterModel}
\usage{
inferStutterModel(
  stutterdata,
  print2pdf = NULL,
  signifLevel = 0.05,
  addValid = FALSE,
  modPrec = FALSE
)
}
\arguments{
\item{stutterdata}{Datatable with columns ("Locus","Type","BLMM","Prop.)}

\item{print2pdf}{Name of pdf file for plots}

\item{signifLevel}{Significance level for keeping BLMM variable}

\item{addValid}{Whether to add Validation figures (PP-plots)}

\item{modPrec}{Whether the model should include precision for BLMM variable}
}
\description{
A function to perform beta-regression analysis
}
\details{
The data table can be stutterdata table obtained from the getStutterData function
}
