% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferStuttRegModel.R
\name{inferStuttRegModel}
\alias{inferStuttRegModel}
\title{inferStuttRegModel}
\usage{
inferStuttRegModel(stutterData, minStuttOccurence = 10,
  print2pdf = NULL, signifLevel = 0.05, addValid = FALSE,
  modPrec = FALSE)
}
\arguments{
\item{stutterData}{Datatable with columns ("Locus","Type","BLMM","Prop.)}

\item{minStuttOccurence}{The number of required observations in order to fit a linear regression with BLMM}

\item{print2pdf}{Name of pdf file for plots}

\item{signifLevel}{Significance level for keeping BLMM variable}

\item{addValid}{Whether to add Validation figures (PP-plots)}

\item{modPrec}{Whether the model should include precision for BLMM variable}
}
\description{
A function to perform beta-regression analysis
}
\details{
The data table can be stutterdata table obtained from the obtainStutterData function
}
