% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMissingTrueAlleles.R
\name{insertMissingTrueAlleles}
\alias{insertMissingTrueAlleles}
\title{insertMissingTrueAlleles}
\usage{
insertMissingTrueAlleles(dat)
}
\arguments{
\item{dat}{Datatable with columns (RefName,SampleName,Locus,Allele,Coverage,Dose)}
}
\description{
Indicate missing alleles of true contributor and insert into table if missing
}
