% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLiki.R
\name{logLiki}
\alias{logLiki}
\title{logLiki}
\usage{
logLiki(mlefit)
}
\arguments{
\item{mlefit}{Fitted object using inferEvidence}
}
\value{
ret A vector with log-likelihood-values for each locus for given model
}
\description{
logLiki returns the likelihood for each markers (based on the MLE results)
}
\author{
Oyvind Bleka
}
