% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeLik.R
\name{optimizeLik}
\alias{optimizeLik}
\title{optimizeLik}
\usage{
optimizeLik(
  logLik,
  phi0,
  maxIter = NULL,
  delta = NULL,
  steptol = NULL,
  iterlim = NULL,
  verbose = TRUE
)
}
\arguments{
\item{logLik}{Objective function to optimize}

\item{phi0}{Start value of parameters (real space)}

\item{maxIter}{Maximum trials of optimization}

\item{delta}{SD of randomly drawed start values}

\item{steptol}{Step tolerance in nlm}

\item{iterlim}{Iteration limit in nlm}

\item{verbose}{Whether to print out progress}
}
\description{
helpfunction to repeat optim steps with random start points
}
