% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMPS.R
\name{plotTopMPS}
\alias{plotTopMPS}
\title{plotTopMPS}
\usage{
plotTopMPS(
  MLEobj,
  DCobj = NULL,
  locYmax = TRUE,
  options = NULL,
  withStutterModel = TRUE,
  returnOnly = FALSE
)
}
\arguments{
\item{MLEobj}{Fitted object using inferEvidence}

\item{DCobj}{An object returned from devonvolve: Must be run with same object as mlefit}

\item{locYmax}{A boolean of whether Y-axis should be same for all markers (FALSE) or not (TRUE this is default)}

\item{options}{A list of possible plot configurations. See comments below}

\item{withStutterModel}{Whether taking the stutter model into account}

\item{returnOnly}{Whether only returning fig (not plotted)}
}
\value{
sub A plotly widget
}
\description{
MPS data visualizer (interactive)
}
\details{
Plots the expected peak heights of the top genotypes. The peak heights for corresponding alleles (one sample) are superimposed.
}
\author{
Oyvind Bleka
}
