% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefit_prediction.R
\name{prefit_prediction}
\alias{prefit_prediction}
\title{prefit_prediction}
\usage{
prefit_prediction(c)
}
\arguments{
\item{c}{The object returned from prepareC_prediction}
}
\value{
Estimated parameters (mean,coefVar,slopePara) per replicate
}
\description{
Fit parameter start values for MLE
}
\details{
Notice that the marker efficiency is scaled with the PHs in order to be taken into account.
}
