% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareC_prediction.R
\name{prepareC_prediction}
\alias{prepareC_prediction}
\title{prepareC_prediction}
\usage{
prepareC_prediction(
  dat,
  hypothesis,
  calibration,
  kit = NULL,
  platform = "MPS",
  model = "GA"
)
}
\arguments{
\item{dat}{A list returned from prepareData_prediction function (includes evididence,reference,frequency data)}

\item{hypothesis}{A list which defines the hypothesis to evaluate. Must contain the following elements:
NOC Number of contributors in model (Mandatory).
condOrder (conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. 
fst The co-ancestry coefficient. Can be a vector (must contain the marker names). Default is 0.
knownRef Specify known non-contributing references from refData (indices). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{calibration}{A list from calibration results (per marker). Must contain the following elements (in following order):
AT Analytical threshold 
markerEff Marker efficiency param
nNoiseParam Parameter for modeling distr of number of noise (Noise model)
noiseSizeParam  Parameter for modeling size of noise alleles (Noise model)
Reg-coeffs (b0,b1,b2) for different stutter types (each element considers a particular stutter type)}

\item{kit}{shortname of kit. Used to model degradation model. Obtained from getMPSkit().}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only consider only some stutter types (BW1,DBW1,FW1)}

\item{model}{Selected model for the signals (read/peak heights): {"GA"=gamma,"NB"=negative binomial}}
}
\value{
ret A list of data input to call the C-code (used in calcLogLikC_prediction)
}
\description{
Helpfunction to prepare data for calling C (used in calcLogLikC_prediction)
}
\details{
Assumes that all PH below threshold are removed. The function builds the data input to the C-code
}
\author{
Oyvind Bleka
}
