% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_prediction.R
\name{prepareData_prediction}
\alias{prepareData_prediction}
\title{prepareData_prediction}
\usage{
prepareData_prediction(samples, refData = NULL, popFreq = NULL,
  minF = NULL, normalize = FALSE, AT = NULL, fillHomGen = TRUE,
  verbose = FALSE, doQ = TRUE)
}
\arguments{
\item{samples}{A list with evidence profiles [[sample]][[locus]]$adata/hdata}

\item{refData}{A list with reference profiles [[reference]][[locus]]...}

\item{popFreq}{A list with population frequencies [[locus]]}

\item{minF}{The freq value included for new alleles (new alleles as potential stutters will have 0). Default NULL is using min.observed in popFreq.}

\item{normalize}{A boolean of whether normalization should be applied or not. Default is FALSE.}

\item{AT}{A detection threshold value or thresholds per marker. NULL ignores filtering.}

\item{fillHomGen}{A boolean of whether to fill in homozygote genotypes given with one allele}

\item{verbose}{Whether to print out if new alleles are imputed}

\item{doQ}{Whether to do Q-assignation of allele outcome (Recommended!)}
}
\value{
Restructured data in list used as input for functions evaluating the liklihood function
}
\description{
Reorganasing data for calculation preparations per marker (simpler structure)
}
\details{
Helpfunction to reorganising data for prepareC input 
Also takes detection thresholds as argument to remove possible alleles below the thresholds.
}
\author{
Oyvind Bleka
}
