% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStutterTypeRule.R
\name{getStutterTypeRule}
\alias{getStutterTypeRule}
\title{getStutterTypeRule}
\usage{
getStutterTypeRule(type, platform = "MPS")
}
\arguments{
\item{type}{Stutter type. Supports (BW1,BW2,FW1,FW2,DBW1,FWBW,BWFW)}

\item{platform}{Platform type {"MPS","CE"}. NOTE: CE has only LUS based stutter types}
}
\description{
Helpfunction to get motifchange for a given stutter type
}
\details{
The nomenlclature for stutters are as follows:
BW1 means that the LUS motif is providing a backward stutter (n-1).
FW1 means that the LUS motif is providing a forward stutter (n+1).
BW2 means that the non-LUS motif is providing a backward stutter (n-1).
etc... prefix "D" means double (n-2 or n+2)
n0 stutters are named as either FWBW or BWFW depending on which occuring first in the sequence (hence orientation matters!)
}
